+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ## [1.7.5] - 2026-06-29
 ### Fixed
 - Fix Android 15 relaunch freeze.
+- Code cleanup
 
 ## [1.7.4] - 2026-06-29
 ### Fixed